backend/requirements.txt
fastapi==0.103.0
uvicorn==0.23.0
httpx==0.24.0
pydantic==2.3.0
python-dotenv==1.0.0
prometheus-client==0.17.0
fastapi-prometheus-instrumentator==6.1.0